refactoring, move Commons methods to relevant classes
