+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/exec/ExecutionServiceImpl.java
@@ -110,6 +110,8 @@ import org.jetbrains.annotations.Nullable;
 
 import static java.util.Collections.singletonList;
 import static org.apache.ignite.internal.processors.query.calcite.CalciteQueryProcessor.FRAMEWORK_CONFIG;
+import static org.apache.ignite.internal.processors.query.calcite.externalize.RelJsonReader.fromJson;
+import static org.apache.ignite.internal.processors.query.calcite.externalize.RelJsonWriter.toJson;
 
 /**
  *
@@ -742,7 +744,7 @@ public class ExecutionServiceImpl<Row> extends AbstractService implements Execut
                             QueryStartRequest req = new QueryStartRequest(
                                 qryId,
                                 pctx.schemaName(),
-                                Commons.toJson(fragment0.root()),
+                                toJson(fragment0.root()),
                                 pctx.topologyVersion(),
                                 fragmentDesc0,
                                 pctx.parameters());
@@ -858,7 +860,7 @@ public class ExecutionServiceImpl<Row> extends AbstractService implements Execut
                 mailboxRegistry(),
                 exchangeService(),
                 failureProcessor())
-                .go(Commons.fromJson(ctx.cluster(), msg.root()));
+                .go(fromJson(ctx, msg.root()));
 
             assert node instanceof Outbox : node;
